+++ iti-react-core/jest.config.js
@@ -1,0 +1,10 @@
+﻿module.exports = {
+    preset: 'ts-jest',
+    globals: {
+        'ts-jest': {
+            diagnostics: {
+                ignoreCodes: ['151001']
+            }
+        }
+    }
+}