+++ iti-react-core/src/Util/ConversionFunctions.ts
@@ -1,0 +1,14 @@
+﻿export function nullToUndefined<T>(x: T | null | undefined): T | undefined {
+    if (x == null) return undefined
+    return x
+}
+
+export function undefinedToNull<T>(x: T | null | undefined): T | null {
+    if (typeof x === 'undefined') return null
+    return x
+}
+
+export function nullToEmpty(s: string | null | undefined): string {
+    if (typeof s !== 'string') return ''
+    return s
+}