+++ iti-react-core/Changelog.md
@@ -19,3 +19,7 @@
     -   Bring in an improved version of preventNonExistentPage
 -   Export Omit type. It can be removed from your .d.ts file and imported from iti-react-core instead
 -   Bring in PhoneNumberUtil from iti-react and export formatPhoneNumber
+
+# 1.12.3
+
+-   \*\*\* Rename `PSEUDO_PROMISE_CANCELLED` to `PSEUDO_PROMISE_CANCELED`